work in progress version 2
swordwt126 works by calculating difficulty every 6 blocks based on a
MTP-6 block range 12-blocks long, and then modifying this new target
twice by a scaled difference to two targets: a fast target, and a slow
target. the fast target is calculated by a weighted-time sum of 21
6-block samples adding up to 126 blocks. The slow target is calculated
by a weighted-time sum of 16 126-block samples adding up to 2016
blocks, where each sample is calculated using a constant-weight sum of
126 blocks recorde every m_height % 126 == 0. the difference of the
slow and fast target are each taken compared to the rolling 12-block
target, and are multiplied by 1/64 and 1/2 respectively, before all
beeing added to the 12-block target to give the new Target Difficulty.
currently all 3 targets use MTP-6, but perhaps this is not necessary
for all but the rolling 12-block target.